reorganize the different methods in new controllers
